Add match percents to scale and chord buttons
Note that this code base is still very exploratory. The underlyding
kamasi library will be modified to make necessary operations easier, and
the data flow will be much cleaner than it is now. Still trying to
figure out exactly how the data should flow, and what each component
should be aware of.